chore: refactor common functions and shopify customers dag code
